- input 엔터 처리 - ContainerBase 구현부 헤더파일로 이동
