+++ include/ftxui/component/component_base.hpp
@@ -6,6 +6,10 @@
 
 #include "ftxui/component/captured_mouse.hpp"  // for CaptureMouse
 #include "ftxui/dom/elements.hpp"              // for Element
+#include "ftxui/component/event.hpp"  // for Event, Event::Tab, Event::TabReverse, Event::ArrowDown, Event::ArrowLeft, Event::ArrowRight, Event::ArrowUp, Event::End, Event::Home, Event::PageDown, Event::PageUp
+#include "ftxui/component/mouse.hpp"  // for Mouse, Mouse::WheelDown, Mouse::WheelUp
+#include "ftxui/dom/elements.hpp"  // for text, Elements, operator|, reflect, Element, hbox, vbox
+#include "ftxui/screen/box.hpp"  // for Box
 
 namespace ftxui {
 
@@ -20,6 +24,11 @@ class Params;
 class ComponentBase;
 using Component = std::shared_ptr<ComponentBase>;
 using Components = std::vector<Component>;
+
+enum ct {
+    component,
+    container,
+};
 
 /// @brief It implement rendering itself as ftxui::Element. It implement
 /// keyboard navigation by responding to ftxui::Event.
@@ -58,6 +67,8 @@ class ComponentBase {
   // root component contains this object.
   virtual Component ActiveChild();
 
+  virtual int GetComponentType() { return ct::component; };
+
   // Return true when the component contains focusable elements.
   // The non focusable Component will be skipped when navigating using the
   // keyboard.
@@ -83,6 +94,233 @@ class ComponentBase {
  private:
   ComponentBase* parent_ = nullptr;
 };
+
+class ContainerBase : public ComponentBase {
+ public:
+  ContainerBase(Components children, int* selector)
+      : selector_(selector ? selector : &selected_) {
+    for (Component& child : children) {
+      Add(std::move(child));
+    }
+  }
+  int GetComponentType() override { return ct::container; }
+
+  // Component override.
+  bool OnEvent(Event event) override {
+    if (event.is_mouse()) {
+      return OnMouseEvent(event);
+    }
+
+    if (!Focused()) {
+      return false;
+    }
+
+    if (ActiveChild() && ActiveChild()->OnEvent(event)) {
+      return true;
+    }
+
+    return EventHandler(event);
+  }
+
+  Component ActiveChild() override {
+    if (children_.empty()) {
+      return nullptr;
+    }
+
+    return children_[static_cast<size_t>(*selector_) % children_.size()];
+  }
+
+  void SetActiveChild(ComponentBase* child) override {
+    for (size_t i = 0; i < children_.size(); ++i) {
+      if (children_[i].get() == child) {
+        *selector_ = static_cast<int>(i);
+        return;
+      }
+    }
+  }
+
+  void MoveSelector(int dir) {
+    for (int i = *selector_ + dir; i >= 0 && i < int(children_.size());
+         i += dir) {
+      if (children_[i]->Focusable()) {
+        *selector_ = i;
+        return;
+      }
+    }
+  }
+
+  void MoveSelectorWrap(int dir) {
+    if (children_.empty()) {
+      return;
+    }
+    for (size_t offset = 1; offset < children_.size(); ++offset) {
+      const size_t i = ((size_t(*selector_ + offset * dir + children_.size())) %
+                        children_.size());
+      if (children_[i]->Focusable()) {
+        *selector_ = int(i);
+        return;
+      }
+    }
+  }
+
+ protected:
+  // Handlers
+  virtual bool EventHandler(Event ) { return false; }  // NOLINT
+  virtual bool OnMouseEvent(Event event) {
+    return ComponentBase::OnEvent(std::move(event));
+  }
+
+  int selected_ = 0;
+  int* selector_ = nullptr;
+
+  
+};
+
+class VerticalContainer : public ContainerBase {
+ public:
+  using ContainerBase::ContainerBase;
+
+  Element Render() override {
+    Elements elements;
+    for (auto& it : children_) {
+      elements.push_back(it->Render());
+    }
+    if (elements.empty()) {
+      return text("Empty container") | reflect(box_);
+    }
+    return vbox(std::move(elements)) | reflect(box_);
+  }
+
+  bool EventHandler(Event event) override {
+    const int old_selected = *selector_;
+    if (event == Event::ArrowUp || event == Event::Character('k')) {
+      MoveSelector(-1);
+    }
+    if (event == Event::ArrowDown || event == Event::Character('j')) {
+      MoveSelector(+1);
+    }
+    if (event == Event::PageUp) {
+      for (int i = 0; i < box_.y_max - box_.y_min; ++i) {
+        MoveSelector(-1);
+      }
+    }
+    if (event == Event::PageDown) {
+      for (int i = 0; i < box_.y_max - box_.y_min; ++i) {
+        MoveSelector(1);
+      }
+    }
+    if (event == Event::Home) {
+      for (size_t i = 0; i < children_.size(); ++i) {
+        MoveSelector(-1);
+      }
+    }
+    if (event == Event::End) {
+      for (size_t i = 0; i < children_.size(); ++i) {
+        MoveSelector(1);
+      }
+    }
+    if (event == Event::Tab) {
+      MoveSelectorWrap(+1);
+    }
+    if (event == Event::TabReverse) {
+      MoveSelectorWrap(-1);
+    }
+
+    *selector_ = std::max(0, std::min(int(children_.size()) - 1, *selector_));
+    return old_selected != *selector_;
+  }
+
+  bool OnMouseEvent(Event event) override {
+    if (ContainerBase::OnMouseEvent(event)) {
+      return true;
+    }
+
+    if (event.mouse().button != Mouse::WheelUp &&
+        event.mouse().button != Mouse::WheelDown) {
+      return false;
+    }
+
+    if (!box_.Contain(event.mouse().x, event.mouse().y)) {
+      return false;
+    }
+
+    if (event.mouse().button == Mouse::WheelUp) {
+      MoveSelector(-1);
+    }
+    if (event.mouse().button == Mouse::WheelDown) {
+      MoveSelector(+1);
+    }
+    *selector_ = std::max(0, std::min(int(children_.size()) - 1, *selector_));
+
+    return true;
+  }
+
+  Box box_;
+};
+
+class HorizontalContainer : public ContainerBase {
+ public:
+  using ContainerBase::ContainerBase;
+
+  Element Render() override {
+    Elements elements;
+    for (auto& it : children_) {
+      elements.push_back(it->Render());
+    }
+    if (elements.empty()) {
+      return text("Empty container");
+    }
+    return hbox(std::move(elements));
+  }
+
+  bool EventHandler(Event event) override {
+    const int old_selected = *selector_;
+    if (event == Event::ArrowLeft || event == Event::Character('h')) {
+      MoveSelector(-1);
+    }
+    if (event == Event::ArrowRight || event == Event::Character('l')) {
+      MoveSelector(+1);
+    }
+    if (event == Event::Tab) {
+      MoveSelectorWrap(+1);
+    }
+    if (event == Event::TabReverse) {
+      MoveSelectorWrap(-1);
+    }
+
+    *selector_ = std::max(0, std::min(int(children_.size()) - 1, *selector_));
+    return old_selected != *selector_;
+  }
+};
+
+class TabContainer : public ContainerBase {
+ public:
+  using ContainerBase::ContainerBase;
+
+  Element Render() override {
+    const Component active_child = ActiveChild();
+    if (active_child) {
+      return active_child->Render();
+    }
+    return text("Empty container");
+  }
+
+  bool Focusable() const override {
+    if (children_.empty()) {
+      return false;
+    }
+    return children_[size_t(*selector_) % children_.size()]->Focusable();
+  }
+
+  bool OnMouseEvent(Event event) override {
+    return ActiveChild() && ActiveChild()->OnEvent(event);
+  }
+};
+
+
+
+
+
 
 }  // namespace ftxui
